Added UpdateIssue and DleteIssue feature with working frontend. (#5)
* 1) Change the type of Id property in User from 'int' to 'string'. Fixed the appropriated models, dtos, commands, queries, handlers, controller actions;
2) Changed sql-scripts for DB - 01-Tables.sql and 02-Sprocs.sql;
3) Added Authentication and Authorization to the Startup class.

* 1) Added the ICurrentUser interface and implemented it in the CurrentUser class;
2) Registered the created service in Startup;
3) Added the [property: JsonPropertyName('sub')] attribute to the Id in CreateUserCommand.

* Fixed primary ctors in all handlers.

* 1) Fixed storage procedure in 02-Sprocs.sql - added directive 'SET NOCOUNT ON';
2) Added ColumnResponseWithIssuesDto;
3) Addded GetAllWithIssuesHandler;
4) Fixed logger.LogInformation in the handlers;
5) Changed GetAllColumns action in the ColumnsController.

* Saved app before changing repo.

* 1) Added feature to move (drug-and-drop) issue in the board - changed appropriated Controller actions, handlers and repo methods.
2) Fixed models and dtos.
3) Fixed queries and commands for handlers.
4) Added migration - 03-SprocsForBusinessLogic to DB.

* Added UpdateIssue and DleteIssue feature with working front

+++ Board.BusinessLogic/Features/ForColumn/Queries/GetColumnByIdQuery.cs
@@ -3,4 +3,4 @@ using MediatR;
 
 namespace Board.BusinessLogic.Features.ForColumn.Queries;
 
-public record GetColumnByIdQuery(int Id) : IRequest<ColumnResponseDto>;
+public record GetColumnByIdQuery(long Id) : IRequest<ColumnResponseDto>;
